+++ pila/integration_test.go
@@ -37,12 +37,13 @@ func TestIntegrationBasic(t *testing.T) {
 		t.Errorf("stack2.Peek is %v, expected %v", stack2.Peek(), nil)
 	}
 
-	var ok = true
-	for ok {
-		var element interface{}
-		element, ok = stack1.Pop()
-		if ok {
+	for {
+		element, _ := stack1.Pop()
+
+		if element != nil {
 			stack2.Push(element)
+		} else {
+			break
 		}
 	}
 
@@ -60,10 +61,10 @@ func TestIntegrationBasic(t *testing.T) {
 		t.Errorf("stack2.Peek is %v, expected %v", stack2.Peek(), interface{}("foo"))
 	}
 
-	if ok = db.RemoveStack(stack1.ID); !ok {
+	if ok := db.RemoveStack(stack1.ID); !ok {
 		t.Errorf("database %s failed on removing stack %s", db.Name, stack1.Name)
 	}
-	if _, ok = db.Stacks[stack1.ID]; ok {
+	if _, ok := db.Stacks[stack1.ID]; ok {
 		t.Errorf("stack1 %s was found in database %s", stack1.Name, db.Name)
 	}