+++ pila/stack.go
@@ -44,13 +44,21 @@ type Stack struct {
 	// when one of these events happens, but it needs to be set by hand.
 	ReadAt time.Time
 
+	// blocked specifies whether a Stack is blocked, so it can be only read, but
+	// not modified.
+	blocked bool
+
 	// dateMu serves as a mutex to lock dates on concurrent
 	// updates in order to avoid race conditions.
 	dateMu sync.Mutex
 
-	// IDMU provides a mutex to handle concurrent reads and
+	// IDMu provides a mutex to handle concurrent reads and
 	// writes on the Stack ID.
 	IDMu sync.RWMutex
+
+	// blockMu provides a mutex to handle concurrent reads and
+	// writes on the Stack Blocked status.
+	blockMu sync.RWMutex
 
 	// base represents the Stack data structure
 	base stack.Stacker
@@ -75,39 +83,82 @@ func NewStackWithBase(name string, t time.Time, base stack.Stacker) *Stack {
 }
 
 // Push an element on top of the Stack.
-func (s *Stack) Push(element interface{}) {
+func (s *Stack) Push(element interface{}) error {
+	if s.Blocked() {
+		return errors.New("Stack is blocked")
+	}
 	s.base.Push(element)
+	return nil
 }
 
 // Pop removes and returns the element on top of the Stack.
-// If the Stack is empty, it returns false.
-func (s *Stack) Pop() (interface{}, bool) {
-	return s.base.Pop()
+// If the Stack was empty, it returns false.
+func (s *Stack) Pop() (interface{}, error) {
+	if s.Blocked() {
+		return nil, errors.New("Stack is blocked")
+	}
+
+	if element, ok := s.base.Pop(); ok {
+		return element, nil
+	}
+	return nil, errors.New("Stack is empty")
 }
 
 // Base bases the Stack on an element, so this becomes
 // the bottommost one of the Stack.
-func (s *Stack) Base(element interface{}) {
+func (s *Stack) Base(element interface{}) error {
+	if s.Blocked() {
+		return errors.New("Stack is blocked")
+	}
+
 	s.base.Base(element)
+	return nil
 }
 
 // Sweep removes and returns the bottommost element of the Stack.
-// If the Stack is empty, it returns false.
-func (s *Stack) Sweep() (interface{}, bool) {
-	return s.base.Sweep()
+// If the Stack is empty or blocked, it returns an error.
+func (s *Stack) Sweep() (interface{}, error) {
+	if s.Blocked() {
+		return nil, errors.New("Stack is blocked")
+	}
+
+	element, ok := s.base.Sweep()
+	if !ok {
+		return nil, errors.New("Stack is empty")
+	}
+
+	return element, nil
 }
 
 // SweepPush removes and returns the bottommost element of the Stack,
 // and pushes an element on top of it, as an atomic operation.
-// If the Stack is empty, it returns false.
-func (s *Stack) SweepPush(element interface{}) (interface{}, bool) {
-	return s.base.SweepPush(element)
+// If the Stack is empty or blocked, it returns an error.
+func (s *Stack) SweepPush(element interface{}) (interface{}, error) {
+	if s.Blocked() {
+		return nil, errors.New("Stack is blocked")
+	}
+
+	element, ok := s.base.SweepPush(element)
+	if !ok {
+		return nil, errors.New("Stack is empty")
+	}
+
+	return element, nil
 }
 
 // Rotate moves the bottommost element of the Stack
-// to the top.
-func (s *Stack) Rotate() bool {
-	return s.base.Rotate()
+// to the top. If the Stack is empty or blocked,
+// it returns an error.
+func (s *Stack) Rotate() error {
+	if s.Blocked() {
+		return errors.New("Stack is blocked")
+	}
+
+	if !s.base.Rotate() {
+		return errors.New("Stack is empty")
+	}
+
+	return nil
 }
 
 // Size returns the size of the Stack.
@@ -126,8 +177,12 @@ func (s *Stack) Peek() interface{} {
 }
 
 // Flush flushes the content of the Stack.
-func (s *Stack) Flush() {
+func (s *Stack) Flush() error {
+	if s.Blocked() {
+		return errors.New("Stack is blocked")
+	}
 	s.base.Flush()
+	return nil
 }
 
 // Update takes a date and updates UpdateAt and ReadAt
@@ -191,11 +246,34 @@ func (s *Stack) Status() StackStatus {
 	status.Name = s.Name
 	status.Size = s.Size()
 	status.Peek = s.Peek()
+	status.Blocked = s.Blocked()
 	status.CreatedAt = s.CreatedAt.Local()
 	status.UpdatedAt = s.UpdatedAt.Local()
 	status.ReadAt = s.ReadAt.Local()
 
 	return status
+}
+
+// Block blocks the Stack.
+func (s *Stack) Block() {
+	s.blockMu.Lock()
+	s.blocked = true
+	s.blockMu.Unlock()
+}
+
+// Unblock unblocks the Stack.
+func (s *Stack) Unblock() {
+	s.blockMu.Lock()
+	s.blocked = false
+	s.blockMu.Unlock()
+}
+
+// Blocked returns true if Stack is blocked.
+func (s *Stack) Blocked() bool {
+	s.blockMu.RLock()
+	defer s.blockMu.RUnlock()
+
+	return s.blocked
 }
 
 // Element represents the payload of a Stack element.