+++ pila/stack_status_test.go
@@ -20,7 +20,7 @@ func TestStackStatusJSON(t *testing.T) {
 	stack.Push([]byte("test"))
 	stack.Update(after)
 
-	expectedStatus := fmt.Sprintf(`{"id":"c8fea3b0-26fd-5ffe-9006-0a71a50ae483","name":"test-stack","peek":"dGVzdA==","size":4,"created_at":"%v","updated_at":"%v","read_at":"%v"}`,
+	expectedStatus := fmt.Sprintf(`{"id":"c8fea3b0-26fd-5ffe-9006-0a71a50ae483","name":"test-stack","peek":"dGVzdA==","size":4,"blocked":false,"created_at":"%v","updated_at":"%v","read_at":"%v"}`,
 		date.Format(now.Local()),
 		date.Format(after.Local()),
 		date.Format(after.Local()))
@@ -36,7 +36,7 @@ func TestStackStatusJSON_Empty(t *testing.T) {
 	stack := NewStack("test-stack", now)
 	stack.Update(now)
 
-	expectedStatus := fmt.Sprintf(`{"id":"c8fea3b0-26fd-5ffe-9006-0a71a50ae483","name":"test-stack","peek":null,"size":0,"created_at":"%v","updated_at":"%v","read_at":"%v"}`,
+	expectedStatus := fmt.Sprintf(`{"id":"c8fea3b0-26fd-5ffe-9006-0a71a50ae483","name":"test-stack","peek":null,"size":0,"blocked":false,"created_at":"%v","updated_at":"%v","read_at":"%v"}`,
 		date.Format(now.Local()),
 		date.Format(now.Local()),
 		date.Format(now.Local()))
@@ -86,13 +86,14 @@ func TestStacksStatusJSON(t *testing.T) {
 	stack2.Push("foo")
 	stack2.Push([]byte("bar"))
 	stack2.Push(999)
+	stack2.Block()
 	stack2.Update(after)
 
 	stacksStatus := StacksStatus{
 		Stacks: []StackStatus{stack1.Status(), stack2.Status()},
 	}
 
-	expectedStatus := fmt.Sprintf(`{"stacks":[{"id":"f6c83edc-f57a-5d0b-a79c-1e4fd365bce1","name":"test-stack-1","peek":"dGVzdA==","size":4,"created_at":"%v","updated_at":"%v","read_at":"%v"},{"id":"7fda1b97-80c9-50f4-86a4-a85d8dd3172d","name":"test-stack-2","peek":999,"size":3,"created_at":"%v","updated_at":"%v","read_at":"%v"}]}`,
+	expectedStatus := fmt.Sprintf(`{"stacks":[{"id":"f6c83edc-f57a-5d0b-a79c-1e4fd365bce1","name":"test-stack-1","peek":"dGVzdA==","size":4,"blocked":false,"created_at":"%v","updated_at":"%v","read_at":"%v"},{"id":"7fda1b97-80c9-50f4-86a4-a85d8dd3172d","name":"test-stack-2","peek":999,"size":3,"blocked":true,"created_at":"%v","updated_at":"%v","read_at":"%v"}]}`,
 		date.Format(now.Local()), date.Format(after.Local()), date.Format(after.Local()),
 		date.Format(now.Local()), date.Format(after.Local()), date.Format(after.Local()))
 	if status, err := stacksStatus.ToJSON(); err != nil {