+++ pila/stack_test.go
@@ -18,6 +18,7 @@ func (s *TestBaseStack) Rotate() bool                                      { ret
 func (s *TestBaseStack) Size() int                                         { return 0 }
 func (s *TestBaseStack) Peek() interface{}                                 { return nil }
 func (s *TestBaseStack) Flush()                                            { return }
+func (s *TestBaseStack) Block() bool                                       { return false }
 
 func TestNewStack(t *testing.T) {
 	now := time.Now()
@@ -101,44 +102,108 @@ func TestStackPush(t *testing.T) {
 	}
 }
 
+func TestStackPush_Blocked(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	stack.Push(1)
+
+	if stack.Size() != 1 {
+		t.Errorf("stack.Size() is %d, expected %d", stack.base.Size(), 1)
+	}
+
+	stack.Block()
+
+	err := stack.Push(2)
+	if err == nil {
+		t.Error("stack should be blocked and non-mutable operations are allowed")
+	}
+
+	if stack.Size() != 1 {
+		t.Errorf("stack.base.Size() is %d, expected %d", stack.base.Size(), 1)
+	}
+}
+
 func TestStackPop(t *testing.T) {
 	stack := NewStack("test-stack", time.Now())
 	stack.Push("test")
+	stack.Push(nil)
 	stack.Push(8)
 
-	element, ok := stack.Pop()
-	if !ok {
+	element, err := stack.Pop()
+	if err != nil {
 		t.Errorf("stack.Pop() not ok")
 	}
 	if element != 8 {
 		t.Errorf("element is %v, expected %v", element, 8)
 	}
-	if stack.Size() != 1 {
-		t.Errorf("stack.Size() is %d, expected %d", stack.Size(), 1)
+	if stack.Size() != 2 {
+		t.Errorf("stack.Size() is %d, expected %d", stack.Size(), 2)
 	}
 }
 
-func TestStackPop_False(t *testing.T) {
+func TestStackPop_Blocked(t *testing.T) {
 	stack := NewStack("test-stack", time.Now())
-	_, ok := stack.Pop()
-	if ok {
-		t.Error("stack.Pop() is ok")
-	}
-}
-
-func TestStackBase(t *testing.T) {
-	stack := NewStack("test-stack", time.Now())
-	stack.Base(1)
+	stack.Push(1)
 
 	if stack.Size() != 1 {
 		t.Errorf("stack.Size() is %d, expected %d", stack.base.Size(), 0)
 	}
 
-	stack.Base(2)
-	stack.Base(struct{ id string }{id: "test"})
+	stack.Block()
+
+	_, err := stack.Pop()
+	if err == nil {
+		t.Error("err is nil, stack should not be blocked", err)
+	}
+
+	if stack.Size() != 1 {
+		t.Errorf("stack.base.Size() is %d, expected %d", stack.base.Size(), 1)
+	}
+}
+
+func TestStackPop_Error(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	peek, err := stack.Pop()
+	if err == nil {
+		t.Errorf("stack.Pop() returned %v, should be empty", peek)
+	}
+}
+
+func TestStackBase(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	_ = stack.Base(1)
+
+	if stack.Size() != 1 {
+		t.Errorf("stack.Size() is %d, expected %d", stack.base.Size(), 0)
+	}
+
+	_ = stack.Base(2)
+	_ = stack.Base(struct{ id string }{id: "test"})
 
 	if stack.Size() != 3 {
 		t.Errorf("stack.Size() is %d, expected %d", stack.Size(), 3)
+	}
+}
+
+func TestStackBase_Blocked(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	err := stack.Base(1)
+	if err != nil {
+		t.Errorf("err is %v, expected nil", err)
+	}
+
+	if stack.Size() != 1 {
+		t.Errorf("stack.Size() is %d, expected %d", stack.base.Size(), 1)
+	}
+
+	stack.Block()
+
+	err = stack.Base(2)
+	if err == nil {
+		t.Error("err is nil, stack should not be blocked", err)
+	}
+
+	if stack.Size() != 1 {
+		t.Errorf("stack.Size() is %d, expected %d", stack.Size(), 1)
 	}
 }
 
@@ -147,9 +212,9 @@ func TestStackSweep(t *testing.T) {
 	stack.Push("test")
 	stack.Push(8)
 
-	element, ok := stack.Sweep()
-	if !ok {
-		t.Errorf("stack.Sweep() not ok")
+	element, err := stack.Sweep()
+	if err != nil {
+		t.Errorf("err is %v, expected nil", err)
 	}
 	if element != "test" {
 		t.Errorf("element is %v, expected %v", element, "test")
@@ -164,9 +229,21 @@ func TestStackSweep(t *testing.T) {
 
 func TestStackSweep_False(t *testing.T) {
 	stack := NewStack("test-stack", time.Now())
-	_, ok := stack.Sweep()
-	if ok {
-		t.Error("stack.Sweep() is ok")
+	if _, err := stack.Sweep(); err == nil {
+		t.Error("err is nil")
+	}
+}
+
+func TestStackSweep_Blocked(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	stack.Push(1)
+
+	stack.Block()
+
+	if _, err := stack.Sweep(); err == nil {
+		t.Error("err is nil")
+	} else if err.Error() != "Stack is blocked" {
+		t.Errorf("err is %v, expected Stack is blocked", err)
 	}
 }
 
@@ -175,9 +252,9 @@ func TestStackSweepPush(t *testing.T) {
 	stack.Push("test")
 	stack.Push(8)
 
-	element, ok := stack.SweepPush("foo")
-	if !ok {
-		t.Errorf("stack.Sweep() not ok")
+	element, err := stack.SweepPush("foo")
+	if err != nil {
+		t.Errorf("err is %v, expected nil", err)
 	}
 	if element != "test" {
 		t.Errorf("element is %v, expected %v", element, "test")
@@ -192,9 +269,21 @@ func TestStackSweepPush(t *testing.T) {
 
 func TestStackSweepPush_False(t *testing.T) {
 	stack := NewStack("test-stack", time.Now())
-	_, ok := stack.SweepPush(8)
-	if ok {
-		t.Error("stack.SweepPush(8) is ok")
+	if _, err := stack.SweepPush(8); err == nil {
+		t.Error("err is nil")
+	}
+}
+
+func TestStackSweepPush_Blocked(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	stack.Push(1)
+
+	stack.Block()
+
+	if _, err := stack.SweepPush(8); err == nil {
+		t.Error("err is nil")
+	} else if err.Error() != "Stack is blocked" {
+		t.Errorf("err is %v, expected Stack is blocked", err)
 	}
 }
 
@@ -204,8 +293,8 @@ func TestStackRotate(t *testing.T) {
 		stack.Push(i)
 	}
 
-	if !stack.Rotate() {
-		t.Errorf("stack.Rotate() is %v, expected %v", false, true)
+	if err := stack.Rotate(); err != nil {
+		t.Errorf("stack.Rotate() is %v, expected nil", err)
 	}
 
 	if stack.Peek() != 0 {
@@ -216,11 +305,11 @@ func TestStackRotate(t *testing.T) {
 	}
 }
 
-func TestStackRotate_False(t *testing.T) {
+func TestStackRotate_Empty(t *testing.T) {
 	stack := NewStack("test-stack", time.Now())
 
-	if stack.Rotate() {
-		t.Errorf("stack.Rotate() is %v, expected %v", true, false)
+	if err := stack.Rotate(); err == nil {
+		t.Errorf("stack.Rotate() is %v, expected %v", err, "Empty")
 	}
 
 	if stack.Peek() != nil {
@@ -228,6 +317,25 @@ func TestStackRotate_False(t *testing.T) {
 	}
 	if stack.Size() != 0 {
 		t.Errorf("stack.Size() is %d, expected %d", stack.Size(), 0)
+	}
+}
+
+func TestStackRotate_Blocked(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	stack.Push(1)
+	stack.Push("abc")
+
+	stack.Block()
+
+	if err := stack.Rotate(); err == nil {
+		t.Errorf("stack.Rotate() is %v, expected %v", err, "Blocked")
+	}
+
+	if stack.Peek() != "abc" {
+		t.Errorf("stack.Peek() is %v, expected %v", stack.Peek(), "abc")
+	}
+	if stack.Size() != 2 {
+		t.Errorf("stack.Size() is %d, expected %d", stack.Size(), 2)
 	}
 }
 
@@ -289,6 +397,26 @@ func TestStackFlush(t *testing.T) {
 	}
 	if stack.Peek() != nil {
 		t.Errorf("stack peek is not nil")
+	}
+}
+
+func TestStackFlush_Blocked(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	stack.Push(1)
+
+	if stack.Size() != 1 {
+		t.Errorf("stack.Size() is %d, expected %d", stack.base.Size(), 0)
+	}
+
+	stack.Block()
+	err := stack.Flush()
+
+	if err == nil {
+		t.Error("stack should be blocked and non-mutable operations are allowed")
+	}
+
+	if stack.Size() != 1 {
+		t.Errorf("stack.base.Size() is %d, expected %d", stack.base.Size(), 1)
 	}
 }
 
@@ -363,6 +491,41 @@ func TestStackSizeToJSON(t *testing.T) {
 	}
 }
 
+func TestStackBlock(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	stack.Push("test")
+	stack.Push(8)
+	stack.Block()
+
+	if !stack.Blocked() {
+		t.Errorf("stack is not blocked")
+	}
+}
+
+func TestStackUnblock(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	stack.Push("test")
+	stack.Push(8)
+	stack.Unblock()
+
+	if stack.Blocked() {
+		t.Errorf("stack is blocked")
+	}
+}
+
+func TestStackBlocked(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	if stack.Blocked() {
+		t.Errorf("stack is blocked")
+	}
+
+	stack.Block()
+
+	if !stack.Blocked() {
+		t.Errorf("stack is not blocked")
+	}
+}
+
 func TestStackRace(t *testing.T) {
 	stack := NewStack("test-stack", time.Now())
 	go func() { stack.Push(1) }()
@@ -389,6 +552,23 @@ func TestStackRace_ID(t *testing.T) {
 	go func() { stack.SetID() }()
 	go func() { _ = stack.UUID() }()
 	go func() { _ = stack.Status() }()
+}
+
+func TestStackRace_Block(t *testing.T) {
+	stack := NewStack("test-stack", time.Now())
+	go func() { stack.Push(1) }()
+	go func() { stack.Pop() }()
+	go func() { stack.Block() }()
+	go func() { stack.Unblock() }()
+	go func() { stack.Update(time.Now()) }()
+	go func() { stack.Size() }()
+	go func() { stack.Unblock() }()
+	go func() { stack.Block() }()
+	go func() { stack.Read(time.Now()) }()
+	go func() { stack.Peek() }()
+	go func() { stack.Block() }()
+	go func() { stack.Unblock() }()
+	go func() { stack.Flush() }()
 }
 
 func TestElementJSON(t *testing.T) {